Reduce false-positives in acquisition

+++ src/board/nap/acq_channel.h
@@ -50,7 +50,7 @@ void nap_acq_load_wr_enable_blocking(void);
 void nap_acq_load_wr_disable_blocking(void);
 void nap_acq_init_wr_params_blocking(s16 carrier_freq);
 void nap_acq_init_wr_disable_blocking(void);
-void nap_acq_corr_rd_blocking(u16 *index, u16 *max, u16 *ave);
+void nap_acq_corr_rd_blocking(u16 *index, u16 *max, float *ave);
 void nap_acq_code_wr_blocking(u8 prn);
 
 #endif  /* SWIFTNAV_ACQ_CHANNEL_H */
